optimization to stop rerender in reviews list
